services/excelupload/requirements.txt
functions-framework==3.*
fastapi
uvicorn
openpyxl
python-multipart
flask-cors